Changed the color scheme and design of the main UI elements.
Inventory, quest, quest entry, and container, now have slightly different colors and the tabbed design was changed. Instead of hard dividers inactive tabs are slightly transparent while the active tab is fully visible.

Added an experimental ring that will teleport the player to the mycelia world, to be used in the tutorial soon.

- EM-103: Fix items that need to be scaled having incorrect offsets in inventory and HUD
- EM-104: Add various mip-map labels to quests and other components