Added randomly generated loot-groves
Loot groves are defined in the tiled map, there is a parent node for identification and then X amount of child nodes which represents where the loot will generate.

There is a rarity system which defines which rarity tear items will spawn, there is a small chance for the rarity to be a step above the current rarity per randomly generated item.

There is a small chance to get more than 1 item

There is a chance to get a container and another smaller chance for that container to have multiple items